Create standalone e2e test image
Upstream-commit: 97f88574613eea15792bafcb81e24619404059b4
Component: cli

+++ components/cli/e2e/stack/remove_test.go
@@ -1,6 +1,7 @@
 package stack
 
 import (
+	"fmt"
 	"strings"
 	"testing"
 
@@ -8,50 +9,69 @@ import (
 	"github.com/gotestyourself/gotestyourself/golden"
 	"github.com/gotestyourself/gotestyourself/icmd"
 	"github.com/gotestyourself/gotestyourself/poll"
+	"github.com/gotestyourself/gotestyourself/skip"
 )
 
 var pollSettings = environment.DefaultPollSettings
 
 func TestRemove(t *testing.T) {
-	stackname := "test-stack-remove"
-	deployFullStack(t, stackname)
-	defer cleanupFullStack(t, stackname)
+	t.Run("Swarm", func(t *testing.T) {
+		testRemove(t, "swarm")
+	})
+	t.Run("Kubernetes", func(t *testing.T) {
+		skip.If(t, !environment.KubernetesEnabled())
 
-	result := icmd.RunCommand("docker", "stack", "rm", stackname)
-
-	result.Assert(t, icmd.Expected{Err: icmd.None})
-	golden.Assert(t, result.Stdout(), "stack-remove-success.golden")
+		testRemove(t, "kubernetes")
+	})
 }
 
-func deployFullStack(t *testing.T, stackname string) {
+func testRemove(t *testing.T, orchestrator string) {
+	stackname := "test-stack-remove-" + orchestrator
+	deployFullStack(t, orchestrator, stackname)
+	defer cleanupFullStack(t, orchestrator, stackname)
+	result := icmd.RunCommand("docker", "--orchestrator", orchestrator,
+		"stack", "rm", stackname)
+	result.Assert(t, icmd.Expected{Err: icmd.None})
+	golden.Assert(t, result.Stdout(),
+		fmt.Sprintf("stack-remove-%s-success.golden", orchestrator))
+}
+
+func deployFullStack(t *testing.T, orchestrator, stackname string) {
 	// TODO: this stack should have full options not minimal options
-	result := icmd.RunCommand("docker", "stack", "deploy",
-		"--compose-file=./testdata/full-stack.yml", stackname)
+	result := icmd.RunCommand("docker", "--orchestrator", orchestrator,
+		"stack", "deploy", "--compose-file=./testdata/full-stack.yml", stackname)
 	result.Assert(t, icmd.Success)
 
-	poll.WaitOn(t, taskCount(stackname, 2), pollSettings)
+	poll.WaitOn(t, taskCount(orchestrator, stackname, 2), pollSettings)
 }
 
-func cleanupFullStack(t *testing.T, stackname string) {
+func cleanupFullStack(t *testing.T, orchestrator, stackname string) {
 	// FIXME(vdemeester) we shouldn't have to do that. it is hidding a race on docker stack rm
-	poll.WaitOn(t, stackRm(stackname), pollSettings)
-	poll.WaitOn(t, taskCount(stackname, 0), pollSettings)
+	poll.WaitOn(t, stackRm(orchestrator, stackname), pollSettings)
+	poll.WaitOn(t, taskCount(orchestrator, stackname, 0), pollSettings)
 }
 
-func stackRm(stackname string) func(t poll.LogT) poll.Result {
+func stackRm(orchestrator, stackname string) func(t poll.LogT) poll.Result {
 	return func(poll.LogT) poll.Result {
-		result := icmd.RunCommand("docker", "stack", "rm", stackname)
+		result := icmd.RunCommand("docker", "--orchestrator", orchestrator, "stack", "rm", stackname)
 		if result.Error != nil {
+			if strings.Contains(result.Stderr(), "not found") {
+				return poll.Success()
+			}
 			return poll.Continue("docker stack rm %s failed : %v", stackname, result.Error)
 		}
 		return poll.Success()
 	}
 }
 
-func taskCount(stackname string, expected int) func(t poll.LogT) poll.Result {
+func taskCount(orchestrator, stackname string, expected int) func(t poll.LogT) poll.Result {
 	return func(poll.LogT) poll.Result {
-		result := icmd.RunCommand(
-			"docker", "stack", "ps", "-f=desired-state=running", stackname)
+		args := []string{"--orchestrator", orchestrator, "stack", "ps", stackname}
+		// FIXME(chris-crone): remove when we support filtering by desired-state on kubernetes
+		if orchestrator == "swarm" {
+			args = append(args, "-f=desired-state=running")
+		}
+		result := icmd.RunCommand("docker", args...)
 		count := lines(result.Stdout()) - 1
 		if count == expected {
 			return poll.Success()
